% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCelltype.R
\name{checkInputs}
\alias{checkInputs}
\title{Check inputs for FindCelltypes function}
\usage{
checkInputs(markers.file, consensus.file, object, filename)
}
\arguments{
\item{markers.file}{Path to the markers file (FindAllMarkers() output)}

\item{consensus.file}{Path to the consensus file}

\item{object}{A Seurat object}

\item{filename}{Character string to name the saved umap plots}
}
\value{
Return markers and consensus files as dataframes
}
\description{
Check inputs for FindCelltypes function
}
