% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCelltype.R
\name{FindCelltype}
\alias{FindCelltype}
\title{Identify cell types based on a user defined consensus markers}
\usage{
FindCelltype(
  markers.file,
  consensus.file,
  object,
  output.dir,
  filename = NULL,
  save.rds = T,
  label = T,
  ...
)
}
\arguments{
\item{markers.file}{Path to the markers file (FindAllMarkers() output)}

\item{consensus.file}{Path to the consensus file}

\item{object}{A Seurat object}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}

\item{filename}{Character string to name the saved umap plots}

\item{save.rds}{Save the final Seurat object in RDS format (default set to True)}

\item{label}{Boolean to display cluster labels on umap plots}

\item{...}{Arguments passed to \code{\link[SeuratObject]{as.sparse}}}
}
\value{
A Seurat Object with cells identities assigned to clusters
}
\description{
Assign cell identities to clusters
}
