% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCelltype.R
\name{getAssignmentsVectors}
\alias{getAssignmentsVectors}
\title{Assign clusters to cell identities from the consensus file provided}
\usage{
getAssignmentsVectors(consensus, markers)
}
\arguments{
\item{consensus}{Consensus dataframe (2 columns per celltype: markers and log-fold)}

\item{markers}{Markers dataframe (FindAllMarkers() output)}
}
\value{
Return two vectors with cell assignments. One using the max number of overlap. The other using the max total FoldChange.
}
\description{
Assign clusters to cell identities from the consensus file provided
}
