% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_object.R
\name{merge_object}
\alias{merge_object}
\title{Merge a list of rds file Seurat object}
\usage{
merge_object(rds.list, output.dir = getwd(), save.rds = TRUE, ...)
}
\arguments{
\item{rds.list}{A vector of rds files paths}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}

\item{save.rds}{Save the final Seurat object in RDS format (default set to True)}

\item{...}{Additional parameters passed to \code{sctransform::vst}}
}
\value{
A merged Seurat Object
}
\description{
Merge a list of rds file Seurat object
}
