% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read10xData.R
\name{Read10xData}
\alias{Read10xData}
\title{Create Seurat Object from sparse data matrices provided by 10X genomics}
\usage{
Read10xData(
  data.dir,
  output.dir = NULL,
  min.cells = 3,
  min.features = 200,
  mt.pattern = "^mt-",
  project.name = "SeuratProject",
  ...
)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv (or features.tsv), and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load
several data directories. If a named vector is given, the cell barcode names
will be prefixed with the name.}

\item{output.dir}{Path to the destination folder of saved files}

\item{min.cells}{Include features detected in at least this many cells. Will
subset the counts matrix as well. To reintroduce excluded features, create a
new object with a lower cutoff.}

\item{min.features}{Include cells where at least this many features are
detected.}

\item{mt.pattern}{Regex pattern of the mitochondrial genes ('^MT-' or '^mt-')}

\item{project.name}{Name of the project/object used for titles in plots}

\item{...}{Arguments passed to \code{\link[SeuratObject]{as.sparse}}}
}
\value{
A Seurat Object and its QC plots
}
\description{
This function create a Seurat object from a 10x output directory by loading
a matrix (matrix.mtx.gz), a barcode file (barcodes.tsv.gz) and a feature file
(features.tsv.gz) from the current directory.
}
