% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10xdata.R
\name{read_10Xdata}
\alias{read_10Xdata}
\title{Create Seurat Object from sparse data matrices provided by 10X genomics}
\usage{
read_10Xdata(
  data.dir,
  output.dir = NULL,
  min.cells,
  min.features,
  mt.pattern,
  project.name
)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv (or features.tsv), and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load
several data directories. If a named vector is given, the cell barcode names
will be prefixed with the name.}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}

\item{mt.pattern}{The regex pattern of the mitochondrial genes ('^MT-' or '^mt-')}
}
\value{
A Seurat Object and its QC plots
}
\description{
This function create a Seurat object from a 10x output directory by loading
a matrix (matrix.mtx.gz), a barcode file (barcodes.tsv.gz) and a feature file
(features.tsv.gz) from the current directory.
}
