% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_integration.R
\name{run_integration}
\alias{run_integration}
\title{Run Seurat Anchors Integration Pipeline}
\usage{
run_integration(
  rds.list,
  output.dir = getwd(),
  resolution = 0.5,
  grouping.var = "orig.ident",
  find.conserved.markers = FALSE,
  ident.1 = NULL,
  ident.2 = NULL,
  save.rds = TRUE,
  ...
)
}
\arguments{
\item{rds.list}{A vector of rds files paths}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}

\item{resolution}{Value of the resolution parameter, use a value above
(below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{save.rds}{Save the final Seurat object in RDS format (default set to True)}

\item{...}{Additional parameters passed to \code{sctransform::vst}}
}
\value{
An integrated Seurat Object along with its visualization plots
}
\description{
This function run the following steps of the Seurat Anchors Integration pipeline :
\itemize{
\item Create Seurat Object
\item QC
\item Normalization
\item Dimensional Reduction
\item Clustering
\item Visualization
\item Differential Gene Expression
}
}
