% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tfactivity.R
\name{run_tfactivity_pipeline}
\alias{run_tfactivity_pipeline}
\title{Run the Seurat Pipeline using TF activity (Dorothea + Viper)}
\usage{
run_tfactivity_pipeline(
  data.dir = NULL,
  output.dir = NULL,
  object = NULL,
  species = "mouse",
  dorothea.confidence = c("A", "B", "C"),
  ...
)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv (or features.tsv), and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load
several data directories. If a named vector is given, the cell barcode names
will be prefixed with the name.}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}

\item{object}{Path to the output directory - where plots and markers file will be saved}

\item{species}{Save the final Seurat object in RDS format (default set to True)}

\item{dorothea.confidence}{To run pipeline on the integrated assay}

\item{save.rds}{Save the final Seurat object in RDS format (default set to True)}

\item{integrated.assay}{To run pipeline on the integrated assay}
}
\value{
A processed Seurat Object along with visualization plots
}
\description{
This function uses Dorothea and Viper to cluster single-cells by TF activity
}
