% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSeurat.R
\name{runCellCycle}
\alias{runCellCycle}
\title{Perform cell cycle scoring}
\usage{
runCellCycle(object, output.dir, no.plot, sctransform)
}
\arguments{
\item{object}{A Seurat object}

\item{output.dir}{Path to the destination folder of saved files}

\item{no.plot}{If set, run the pipeline without saving the plots}

\item{sctransform}{If set, use SCTransform normalization}
}
\value{
A Seurat object with Cell Cycle phase for each cells
}
\description{
Perform cell cycle scoring
}
