% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSeurat.R
\name{runClustering}
\alias{runClustering}
\title{Perform clustering on Seurat object}
\usage{
runClustering(
  object,
  output.dir,
  ndims,
  dims,
  resolution,
  no.plot,
  sctransform
)
}
\arguments{
\item{object}{A Seurat object}

\item{output.dir}{Path to the destination folder of saved files}

\item{ndims}{Number of dimensions to plot standard deviation for}

\item{dims}{Which dimensions to use as input features, used only if
\code{features} is NULL}

\item{resolution}{Value of the resolution parameter, use a value above
(below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{no.plot}{If set, run the pipeline without saving the plots}

\item{sctransform}{If set, use SCTransform normalization}
}
\value{
A Seurat object with cells clustered and some cluster visualizations
}
\description{
Perform clustering on Seurat object
}
