% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTfPipeline.R
\name{RunTf}
\alias{RunTf}
\title{Create Dorothea Assay with TF activity computed using Viper}
\usage{
RunTf(
  data.dir = NULL,
  object = NULL,
  species = "mouse",
  dorothea.confidence = c("A", "B", "C"),
  ...
)
}
\arguments{
\item{data.dir}{Directory containing the matrix.mtx, genes.tsv (or features.tsv), and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load
several data directories. If a named vector is given, the cell barcode names
will be prefixed with the name.}

\item{object}{A Seurat object}

\item{species}{Mouse or human}

\item{dorothea.confidence}{Confidence levels of regulons}

\item{...}{Arguments passed to \code{\link[SeuratObject]{as.sparse}}}
}
\value{
A processed Seurat Object along with visualization plots based on TF activity
}
\description{
Create Dorothea Assay with TF activity computed using Viper
}
