% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunTfPipeline.R
\name{RunTfPipeline}
\alias{RunTfPipeline}
\title{Run the Seurat Pipeline using TF activity (Dorothea + Viper)}
\usage{
RunTfPipeline(
  data.dir = NULL,
  object = NULL,
  species = "mouse",
  dorothea.confidence = c("A", "B", "C"),
  ...
)
}
\arguments{
\item{object}{A Seurat object}

\item{species}{Mouse or human}

\item{dorothea.confidence}{Confidence levels of regulons}
}
\value{
A processed Seurat Object along with visualization plots
}
\description{
This function uses Dorothea and Viper to cluster single-cells by TF activity
}
