% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save output plot in png format}
\usage{
save_plot(filename, plot, output.dir)
}
\arguments{
\item{filename}{The filename under which to save the plot}

\item{plot}{The plot to save}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}
}
\value{
The plot saved as a png image
}
\description{
Save output plot in png format
}
