% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavePlot.R
\name{SavePlot}
\alias{SavePlot}
\title{Save output plot in png format}
\usage{
SavePlot(filename, plot, output.dir)
}
\arguments{
\item{filename}{The filename under which to save the plot}

\item{plot}{The plot to save}

\item{output.dir}{Path to the destination folder of saved files}
}
\value{
The plot saved as a png image
}
\description{
Save output plot in png format
}
