% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergeObject.R
\name{MergeObject}
\alias{MergeObject}
\title{Merge a list of rds file Seurat object}
\usage{
MergeObject(
  rds.list,
  obj.list = NULL,
  output.dir = getwd(),
  save.rds = TRUE,
  ...
)
}
\arguments{
\item{rds.list}{A vector of rds files paths}

\item{obj.list}{A vector of Seurat objects}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}

\item{save.rds}{Save the final Seurat object in RDS format (default set to True)}

\item{...}{Arguments passed to \code{\link[SeuratObject]{as.sparse}}}
}
\value{
A merged Seurat Object
}
\description{
Merge a list of rds file Seurat object
}
