% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSeurat.R
\name{runCellCycle}
\alias{runCellCycle}
\title{Perform cell cycle scoring}
\usage{
runCellCycle(object, output.dir, no.plot)
}
\arguments{
\item{object}{A Seurat object}

\item{output.dir}{Path to the output directory - where plots and markers file will be saved}

\item{no.plot}{Run the pipeline without saving the plots (default set to False)}
}
\value{
A Seurat object with Cell Cycle phase for each cells
}
\description{
Perform cell cycle scoring
}
